refactor: extract insert group actions
